Change card appearance and add menu
