+++ components/Card/Card.module.scss
@@ -9,10 +9,10 @@
   /* UI Background/Default */
 
   width: 332px;
-  height: 168px;
+  height: 100px;
 
   min-width: 332px;
-  max-height: 168px;
+  max-height: 100px;
 
   background: #ffffff;
   border-radius: 4px;
@@ -35,7 +35,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  max-height: 70px;
+  max-height: 40px;
   overflow: hidden;
 }
 
@@ -56,11 +56,23 @@
   }
 }
 
+.topSection {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  overflow: hidden;
+}
+
+.favIconContainer {
+  align-self: center;
+  margin-right: 16px;
+}
+
 .vsmTitleContainer {
-  margin: 8px 16px 16px;
   flex: 1;
   overflow: hidden;
   text-overflow: ellipsis;
+  margin: 8px 16px 16px;
 }
 
 .vsmTitle {
