+++ components/Card/Card.tsx
@@ -1,23 +1,51 @@
 import moment from "moment";
-import React from "react";
+import React, { useState } from "react";
 import Link from "next/link";
 import styles from "./Card.module.scss";
 import { UserDots } from "../UserDots";
 import { vsmProject } from "../../interfaces/VsmProject";
+import { favorite_outlined, favorite_filled } from "@equinor/eds-icons";
+import { Icon } from "@equinor/eds-core-react";
 
 export function VSMCard(props: { vsm: vsmProject }): JSX.Element {
   const { userIdentity: createdBy } = props.vsm.created;
+  const [isHighlighted, setIsHighlighted] = useState(false);
+  const [isFavourite, setIsFavourite] = useState(false);
 
   return (
     <Link href={`/projects/${props.vsm.vsmProjectID}`}>
       <div className={styles.card}>
-        <div className={styles.vsmTitleContainer}>
-          <h1 className={styles.vsmTitle}>
-            {props.vsm.name || "Untitled VSM"}
-          </h1>
+        <div className={styles.topSection}>
+          <div className={styles.vsmTitleContainer}>
+            <h1 className={styles.vsmTitle}>
+              {props.vsm.name || "Untitled VSM"}
+            </h1>
+          </div>
+          <div
+            className={styles.favIconContainer}
+            onMouseOver={(e) => {
+              e.stopPropagation();
+              setIsHighlighted(true);
+            }}
+            onMouseLeave={(e) => {
+              e.stopPropagation();
+              setIsHighlighted(false);
+            }}
+            onClick={(e) => {
+              e.stopPropagation();
+              setIsFavourite((previous) => !previous);
+            }}
+          >
+            {isFavourite ? (
+              <Icon color="#ff1243" data={favorite_filled} />
+            ) : isHighlighted ? (
+              <Icon color="#DADADA" data={favorite_filled} />
+            ) : (
+              <Icon color="#DADADA" data={favorite_outlined} />
+            )}
+          </div>
         </div>
         <div>
-          <hr style={{ opacity: 0.1 }} />
           <div className={styles.bottomSection}>
             {!!props.vsm.lastUpdated && (
               <p className={styles.vsmLabel}>
