+++ components/Card/LoadingVSMCard.module.scss
@@ -1,6 +1,6 @@
 .card {
   width: 332px;
-  height: 168px;
+  height: 100px;
 
   background: #ffffff;
   border-radius: 4px;
