+++ pages/projects/Projects.module.scss
@@ -34,10 +34,10 @@
   align-items: center;
 
   width: 332px;
-  height: 168px;
+  height: 100px;
 
   min-width: 332px;
-  max-height: 168px;
+  max-height: 100px;
 
   border-radius: 4px;
